+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next Up
 
 - Resolve some character update bugs introduced with v5.1.15.
+- Make monster and item names use @lookup in when hiding text from players.
 
 # 5.1.16